levels are loaded

+++ lib_stage_screen/stage.h
@@ -25,8 +25,8 @@ public:
 
     void render(Window *window) override;
 
-    void setPosition(int x, int y) override;
-    void shiftPosition(int dx, int dy) override{};
+    void setPosition(float x, float y) override;
+    void shiftPosition(float dx, float dy) override{};
 
     static bool onMouseMove(sf::Event &event, void *);
     static bool onMousePress(sf::Event &event, void *);
